code fixes, refactor for tags command group

+++ src/WinTenBot/Zizi.Bot/Handlers/Commands/Tags/UntagCommand.cs
@@ -2,65 +2,68 @@
 using System.Threading.Tasks;
 using Serilog;
 using Telegram.Bot.Framework.Abstractions;
-using Zizi.Bot.Telegram;
 using Zizi.Bot.Services;
+using Zizi.Bot.Services.Datas;
+using Zizi.Bot.Telegram;
+using Zizi.Core.Utils;
+using Zizi.Core.Utils.Text;
 
 namespace Zizi.Bot.Handlers.Commands.Tags
 {
     public class UntagCommand : CommandBase
     {
-        private TagsService _tagsService;
-        private TelegramService _telegramService;
+        private readonly TagsService _tagsService;
+        private readonly TelegramService _telegramService;
 
-        public UntagCommand(TagsService tagsService)
+        public UntagCommand(TagsService tagsService, TelegramService telegramService)
         {
             _tagsService = tagsService;
+            _telegramService = telegramService;
         }
 
         public override async Task HandleAsync(IUpdateContext context, UpdateDelegate next, string[] args,
             CancellationToken cancellationToken)
         {
-            _telegramService = new TelegramService(context);
+            await _telegramService.AddUpdateContext(context);
+
             var msg = _telegramService.Message;
 
             var isSudoer = _telegramService.IsSudoer();
-            var isAdmin = await _telegramService.IsAdminOrPrivateChat()
-                .ConfigureAwait(false);
-            var tagVal = args[0];
+            var isAdmin = _telegramService.IsAdminOrPrivateChat();
+            var tagVal = _telegramService.MessageTextParts.ValueOfIndex(1);
             var chatId = _telegramService.Message.Chat.Id;
             var sendText = "Hanya admin yang bisa membuat Tag.";
 
             if (!isAdmin && !isSudoer)
             {
-                await _telegramService.SendTextAsync(sendText)
-                    .ConfigureAwait(false);
+                await _telegramService.SendTextAsync(sendText);
                 Log.Information("This User is not Admin or Sudo!");
                 return;
             }
 
-            await _telegramService.SendTextAsync("Memeriksa..")
-                .ConfigureAwait(false);
-            var isExist = await _tagsService.IsExist(chatId, tagVal)
-                .ConfigureAwait(false);
+            if (tagVal.IsNullOrEmpty())
+            {
+                await _telegramService.SendTextAsync("Tag apa yg mau di hapus?");
+                return;
+            }
+
+            await _telegramService.SendTextAsync("Memeriksa..");
+            var isExist = await _tagsService.IsExist(chatId, tagVal);
             if (isExist)
             {
-                Log.Information($"Sedang menghapus tag {tagVal}");
-                var unTag = await _tagsService.DeleteTag(chatId, tagVal)
-                    .ConfigureAwait(false);
+                Log.Information("Sedang menghapus tag {TagVal}", tagVal);
+                var unTag = await _tagsService.DeleteTag(chatId, tagVal);
                 if (unTag)
                 {
                     sendText = $"Hapus tag {tagVal} berhasil";
                 }
 
-                await _telegramService.EditAsync(sendText)
-                    .ConfigureAwait(false);
-                await _tagsService.UpdateCacheAsync(msg)
-                    .ConfigureAwait(false);
-                return;
+                await _telegramService.EditAsync(sendText);
+                await _tagsService.UpdateCacheAsync(chatId);
             }
             else
             {
-                sendText = $"Tag {tagVal} tidak di temukan";
+                await _telegramService.EditAsync( $"Tag {tagVal} tidak di temukan");
             }
         }
     }
